Repository contains a utility that is build to make dakota's SOGA method operate with AI capabilities.
The implementation first modifies dakota's LibraryEnvironment class and
adds a function that can reset the topLevelMethod iterator rep to our
implementation of JEGAOptimizer. This is done to avoid making changes to
dakota and its parser. Dakota's dlsolver capability does not work and is
not fully tested by Sandia.

Next, we copy JEGAOptimizer instead of inheriting it. Inheritance stops us
from using some of the crutial (and required) objects from JEGAOptimizer,
hence a copy is preferred. TODO: Maybe someone else can find a better solution.

In MyJEGAOPtimizer, we first pass the iterated model to the Driver subclass.
This way we can switch the driver script (or executable) before each iteration/
evaluation. We also override PerformIterations and PerformNextIteration methods
to incorporate AI capabilities into SOGA.

Note: To our driver script with always pass the current design's neighbors, and
a yes/no string that will indicate which fluid solver to use. The number of
neighbors is defined in our input file. This way the driver script can handle
switching of fluid solver's and get design id's from which it should interpolate.
In future, we can utilize PluginParallelDirectInterface class to incorporate
m2c and aero-s with gam2c.

Working on IoData and callback utility that will set the dakota data containers.